% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dosresmeta-package.R
\docType{data}
\name{alcohol_crc}
\alias{alcohol_crc}
\title{Eight published studies on the relation between alcohol intake and colon-rectal cancer.}
\format{A data frame with 48 observations on the following 7 variables:
\tabular{ll}{
\code{id} \tab label for author's names (id variable).\cr
\code{type} \tab code for study design.\cr
\code{dose} \tab assigned dose level.\cr
\code{cases} \tab number of cases for each exposure level.\cr
\code{peryears} \tab amount of person-time for each exposure level.\cr
\code{logrr} \tab natural logarithm of adjusted "relative risks".\cr
\code{se} \tab standard errornatural for the logarithm of adjusted "relative risks".\cr
}}
\description{
The dataset reports the summarized dose-response results from eight prospective
#' studies on the relation between alcohol intake and colon-rectal risk (Orsini 2012).
}
\author{
Alessio Crippa, <\email{alessio.crippa@ki.se}>
}
\references{
Orsini, N., Li, R., Wolk, A., Khudyakov, P., Spiegelman, D. (2012). Meta-analysis for linear and nonlinear dose-response relations: examples,
an evaluation of approximations, and software. American journal of epidemiology, 175(1), 66-73.
}
\keyword{data}

