% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{alcohol_cvd}
\alias{alcohol_cvd}
\title{Six published studies on the relation between alcohol intake and cardiovascular disease risk}
\format{
A data frame with 25 observations on the following 8 variables:
\tabular{ll}{
\code{id} \tab id of the studies included in the analysis.\cr
\code{author} \tab names of the first author of the study.\cr
\code{type} \tab code for study design.\cr
\code{dose} \tab assigned dose levels.\cr
\code{cases} \tab number of cases for each exposure level.\cr
\code{n} \tab total number of subjects for each exposure level.\cr
\code{logrr} \tab natural logarithm of the adjusted "relative risks".\cr
\code{se} \tab standard error for the logarithm of the adjusted "relative risks".\cr
}
}
\description{
The dataset reports the summarized dose-response results from six observational
studies on the relation between alcohol intake and vascular disease risk. Four are case-control studies,
two prospective (cumulative-incidence data).
}
\references{
Liu, Q., Cook, N. R., Bergstrom, A., Hsieh, C. C. (2009). A two-stage hierarchical regression model for meta-analysis of epidemiologic nonlinear
dose-response data. Computational Statistics & Data Analysis, 53(12), 4157-4167.
}
\author{
Alessio Crippa, <\email{alessio.crippa@ki.se}>
}
\keyword{data}
