% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{alcohol_esoph}
\alias{alcohol_esoph}
\title{Fourteen case-control studies on the relation between alcohol consumption and esophageal cancer}
\format{A data frame with 63 observations on the following 8 variables:
\tabular{ll}{
\code{id} \tab id of the studies included in the analysis.\cr
\code{author} \tab names of the first author.\cr
\code{type} \tab code for study design.\cr
\code{cases} \tab number of cases for each exposure level.\cr
\code{n} \tab total number of subjects for each exposure level.\cr
\code{dose} \tab assigned dose levels.\cr
\code{logrr} \tab natural logarithm of the adjusted odds ratio.\cr
\code{se} \tab standard error for the logarithm of the adjusted odds ratio\cr
}}
\description{
The dataset reports the summarized dose-response results from fourteen case-control studies on the relation between alcohol consumption
and esophageal squamous cell carcinoma.
}
\references{
Rota M, Bellocco R, Scotti L, Tramacere I, Jenab M, Corrao G, La Vecchia C, Boffetta P, 
Bagnardi V. Random-effects meta-regression models for studying nonlinear dose-response 
relationship, with an application to alcohol and esophageal squamous cell carcinoma. 
Statistics in medicine. 2010 Nov 20;29(26):2679-87.
}
\author{
Alessio Crippa, <\email{alessio.crippa@ki.se}>
}
\keyword{data}
