% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{alcohol_lc}
\alias{alcohol_lc}
\title{Four published studies on the relation between alcohol intake and lung cancer.}
\format{A data frame with 20? observations on the following 7 variables:
\tabular{ll}{
\code{id} \tab label for author's names (id variable).\cr
\code{type} \tab code for study design.\cr
\code{dose} \tab assigned dose levels.\cr
\code{cases} \tab number of cases for each exposure level.\cr
\code{peryears} \tab amount of person-time for each exposure level.\cr
\code{logrr} \tab natural logarithm of the adjusted "relative risks".\cr
\code{se} \tab standard error for the logarithm of the adjusted "relative risks".\cr
}}
\description{
The dataset reports the summarized dose-response results from four prospective
studies on the relation between alcohol intake and lunger cancer.
}
\references{
Orsini, N., Li, R., Wolk, A., Khudyakov, P., Spiegelman, D. (2012). Meta-analysis for linear and nonlinear dose-response relations: examples,
an evaluation of approximations, and software. American journal of epidemiology, 175(1), 66-73.
}
\author{
Alessio Crippa, <\email{alessio.crippa@ki.se}>
}
\keyword{data}
