% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{ari}
\alias{ari}
\title{Five clinical trials on the relation between aripiprazole and schizophrenia}
\format{
A data frame with 18 observations on the following 6 variables:
\tabular{ll}{
\code{id} \tab id of the studies included in the analysis.\cr
\code{author} \tab names of the first author of the studies.\cr
\code{dose} \tab assigned dose level of aripiprazole (0 for placebo group).\cr
\code{y} \tab outcome variable: change in PANNS score after and before treatment.\cr
\code{sd} \tab standard deviation of y for each exposure level.\cr
\code{n} \tab total number of subjects for each exposure level.\cr
}
}
\description{
The dataset reports the summarized dose-response results from five clinical trials on the relation between different levels of aripiprazole
and severety of schizophrenia measured usign the PANSS medical score.
}
\references{
Crippa, A., Orsini, N. Dose-response meta-analysis of differences in means. BMC medical research methodology. 2016 Aug 2;16(1):91.
}
\author{
Alessio Crippa, <\email{alessio.crippa@ki.se}>
}
\keyword{data}
