% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{bmi_rc}
\alias{bmi_rc}
\title{Four case-control studies on the relation between Body Mass Index and renal cell cancer}
\format{
A data frame with 33 observations on the following 13 variables:
\tabular{ll}{
\code{id} \tab id of the studies included in the analysis.\cr
\code{author} \tab names of the first author and year of publication.\cr
\code{type} \tab code for study design.\cr
\code{interval} \tab intervals for the categories of bmi.\cr
\code{bmi} \tab assigned bmi levels.\cr
\code{case} \tab number of cases for each exposure level.\cr
\code{control} \tab number of controls for each exposure level.\cr
\code{n} \tab total number of subjects for each exposure level.\cr
\code{or} \tab adjusted odds ratios for each exposure level.\cr
\code{lb_or} \tab lower bound for the confidence limits of the adjusted odds ratios.\cr
\code{ub_or} \tab upper bound for the confidence limits of the adjusted odds ratios.\cr
\code{logor} \tab natural logarithm of the adjusted odds ratios.\cr
\code{se_logor} \tab standard error for the logarithm of the adjusted odds ratios.\cr
}
}
\description{
The dataset reports the summarized dose-response results from four cases-control studies on the relation
 Body Mass Index and renal cell cancer
}
\references{
Liu, Q., Cook, N. R., Bergstrom, A., Hsieh, C. C. (2009). A two-stage hierarchical regression model for meta-analysis of epidemiologic nonlinear
dose-response data. Computational Statistics & Data Analysis, 53(12), 4157-4167.
}
\author{
Alessio Crippa, <\email{alessio.crippa@ki.se}>
}
\keyword{data}
