\name{dosresmeta-package}
\alias{dosresmeta-package}
\docType{package}
\title{
Performing dose-response meta-analysis  
}
\description{
It allows to estimate a dose-response relation from single and multiple summarized data. The trend estimation takes into account the correlationamong log relative risks and use it to efficiently estimate the dose-response relation for each study included in the anlysis. The results are then combined according to principles of multivariate random effects meta-analysis to obtain a pooled functional relation.
}
\details{
\tabular{ll}{
Package: \tab dosresmeta\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2013-09-15\cr
License: \tab GPL-2\cr
}
}
\author{Alessio Crippa

Maintainer: Alessio Crippa <alessio.crippa@ki.se>
}
\references{
Orsini, N., Bellocco, R., & Greenland, S. (2006). Generalized least squares for trend estimation of summarized dose-response data. Stata Journal, 6(1), 40.

Gasparrini, A., Armstrong, B., & Kenward, M. G. (2012). Multivariate meta-analysis for non-linear and other multi-parameter associations. Statistics in Medicine, 31(29), 3821-3839.
}
\keyword{ package }
\seealso{
	\code{\link{mvmeta}}
}