% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{fish_ra}
\alias{fish_ra}
\title{Six studies on the relation between fish consumption and rheumatoid arthritis risk}
\format{A data frame with 22 observations on the following 12 variables:
\tabular{ll}{
\code{id} \tab id of the studies included in the analysis.\cr
\code{author} \tab names of the first author.\cr
\code{year} \tab year of publication.\cr
\code{type} \tab code for study design.\cr
\code{cases} \tab number of cases for each exposure level.\cr
\code{n} \tab total number of subjects (type = "ci") or person-years (type = "ir") for each exposure level.\cr
\code{dose} \tab assigned dose levels.\cr
\code{rr} \tab adjusted risk estimates.\cr
\code{lrr} \tab lower bound for the confidence limits of the adjusted risk estimates.\cr
\code{urr} \tab upper bound for the confidence limits of the adjusted risk estimates.\cr
\code{logrr} \tab natural logarithm of the adjusted odds ratio.\cr
\code{se} \tab standard error for the logarithm of the adjusted odds ratio\cr
}}
\description{
The dataset reports the summarized dose-response results from six studies on the relation between fish consumption
and rheumatoid arthritis risk
}
\references{
Di Giuseppe D, Crippa A, Orsini N, Wolk A. Fish consumption and risk of rheumatoid 
arthritis: a dose-response meta-analysis. Arthritis research & therapy. 
2014 Sep 30;16(5):446.
}
\author{
Alessio Crippa, <\email{alessio.crippa@ki.se}>
}
\keyword{data}
