% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{milk_mort}
\alias{milk_mort}
\title{Eleven prospective studies on the relation between milk consumption and all-cause mortality}
\format{
A data frame with 50 observations on the following 12 variables:
\tabular{ll}{
\code{id} \tab id of the studies included in the analysis.\cr
\code{author} \tab names of the first author.\cr
\code{year} \tab year of publication.\cr
\code{type} \tab code for study design.\cr
\code{dose} \tab assigned dose levels.\cr
\code{cases} \tab number of cases for each exposure level.\cr
\code{n} \tab total number of subjects (type = "ci") or person-years (type = "ir") for each exposure level.\cr
\code{rr} \tab adjusted risk estimates.\cr
\code{lb} \tab lower bound for the confidence limits of the adjusted risk estimates.\cr
\code{ub} \tab upper bound for the confidence limits of the adjusted risk estimates.\cr
\code{logrr} \tab natural logarithm of the adjusted risk estimates.\cr
\code{se} \tab standard error for the logarithm of the adjusted risk estimates.\cr
}
}
\description{
The dataset reports the summarized dose-response results from eleven prospective studies on the relation between milk consumption
and all-cause mortality.
}
\references{
Larsson SC, Crippa A, Orsini N, Wolk A, Michaelsson K. Milk consumption and mortality from all causes, 
cardiovascular disease, and cancer: a systematic review and meta-analysis. Nutrients. 2015 Sep 11;7(9):7749-63.
}
\author{
Alessio Crippa, <\email{alessio.crippa@ki.se}>
}
\keyword{data}
