% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{milk_ov}
\alias{milk_ov}
\title{Nine studies on the relation between milk consumption and ovarian cancer}
\format{A data frame with 37 observations on the following 12 variables:
\tabular{ll}{
\code{id} \tab id of the studies included in the analysis.\cr
\code{author} \tab names of the first author.\cr
\code{year} \tab year of publication.\cr
\code{type} \tab code for study design.\cr
\code{dose} \tab assigned dose levels.\cr
\code{case} \tab number of cases for each exposure level.\cr
\code{n} \tab total number of subjects (type = "ir" or "cc") or person-years (type = "ir") for each exposure level.\cr
\code{rr} \tab adjusted risk estimates.\cr
\code{lb} \tab lower bound for the confidence limits of the adjusted risk estimates.\cr
\code{ub} \tab upper bound for the confidence limits of the adjusted risk estimates.\cr
\code{logrr} \tab natural logarithm of the adjusted risk estimates.\cr
\code{se} \tab standard error for the logarithm of the adjusted risk estimates.\cr
}}
\description{
The dataset reports the summarized dose-response results from nine studies on the relation between milk consumption
and ovarian cancer.
}
\references{
Larsson, S. C., N. Orsini, and A. Wolk. 2005. Milk, milk products and lactose intake and ovarian cancer risk: A meta-analysis of epidemiological studies. 
International Journal of Cancer 118: 431-441.

Greenland, S.,  Longnecker, M. P. (1992). Methods for trend estimation from summarized dose-response data, with applications to meta-analysis.
American journal of epidemiology, 135(11), 1301-1309.
}
\author{
Alessio Crippa, <\email{alessio.crippa@ki.se}>
}
\keyword{data}
