% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{process_bc}
\alias{process_bc}
\title{Ten studies on the relation between processed meat and bladder cancer}
\format{
A data frame with 73 observations on the following 15 variables:
\tabular{ll}{
\code{id} \tab id of the studies included in the analysis.\cr
\code{author} \tab names of the first author.\cr
\code{year} \tab year of publication.\cr
\code{type} \tab code for study design.\cr
\code{dose0} \tab original assigned dose levels, with unit of measurement defined in the "unit" column.\cr
\code{dose} \tab assigned dose levels (converted (if needed) in gm/day).\cr
\code{cases} \tab number of cases for each exposure level.\cr
\code{n} \tab total number of subjects (type = "ci" or "cc") or person-years (type = "ir") for each exposure level.\cr
\code{rr} \tab adjusted risk estimates.\cr
\code{lb} \tab lower bound for the confidence limits of the adjusted risk estimates.\cr
\code{ub} \tab upper bound for the confidence limits of the adjusted risk estimates.\cr
\code{logrr} \tab natural logarithm of the adjusted risk estimates.\cr
\code{se} \tab standard error for the logarithm of the adjusted risk estimates.\cr
\code{area} \tab geographical area of the published study.\cr
\code{unit} \tab unit of measurement for red meat consumption (for dose0).\cr
}
}
\description{
The dataset reports the summarized dose-response results from ten studies on the relation between processed meat consumption
and bladder cancer.
}
\references{
Crippa A, Larsson SC, Discacciati A, Wolk A, Orsini N. Red and processed meat consumption and 
risk of bladder cancer: a dose-response meta-analysis of epidemiological studies. 
European journal of nutrition. 2016 Dec 22:1-3.
}
\author{
Alessio Crippa, <\email{alessio.crippa@ki.se}>
}
\keyword{data}
