% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{coffee_stroke}
\alias{coffee_stroke}
\title{Eleven prospective studies on the relation between coffee consumption and stroke risk}
\format{A data frame with 68 observations on the following 12 variables:
\tabular{ll}{
\code{id} \tab id of the studies included in the analysis.\cr
\code{author} \tab names of the first author of the studies.\cr
\code{type} \tab code for study design.\cr
\code{dose} \tab assigned dose levels.\cr
\code{cases} \tab number of cases for each exposure level.\cr
\code{n} \tab total number of subjects (type = "ci") or person-years (type = "ir") for each exposure level.\cr
\code{rr} \tab adjusted risk estimates for each exposure level.\cr
\code{lb} \tab lower bound for the confidence limits of the adjusted risk estimates.\cr
\code{ub} \tab upper bound for the confidence limits of the adjusted risk estimates.\cr
\code{logrr} \tab natural logarithm of the adjusted risk estimates.\cr
\code{se} \tab standard error for the logarithm of the adjusted risk estimates.\cr
\code{nordic} \tab indicator variable for the study to be conducted in the nordic countries (1 = yes).\cr
}}
\description{
The dataset reports the summarized dose-response results from eleven prospective studies on the relation between coffee consumption
and risk of stroke.
}
\references{
Larsson, S. C., Orsini, N. (2011). Coffee consumption and risk of stroke: a dose-response 
meta-analysis of prospective studies. American journal of epidemiology, 174(9), 993-1001.
}
\author{
Alessio Crippa, <\email{alessio.crippa@ki.se}>
}
\keyword{data}
