\name{dosresmeta-package}
\alias{dosresmeta-package}
\docType{package}
\title{
Performing dose-response meta-analysis  
}
\description{
It allows to estimate a dose-response relation from single and multiple summarized data. The trend estimation takes into account the correlation
		among log relative risks and use it to efficiently estimate the dose-response relation for each study. The results are then combined according to principles
		of multivariate random effects meta-analysis to obtain a pooled functional relation.
}
\details{
\tabular{ll}{
Package: \tab dosresmeta\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-08-20\cr
License: \tab GPL-2\cr
}}


\author{Alessio Crippa

Maintainer: Alessio Crippa <alessio.crippa@ki.se>
}
\references{
Orsini, N., Bellocco, R., & Greenland, S. (2006). Generalized least squares for trend estimation of summarized dose-response data. Stata Journal, 6(1), 40.
}
\keyword{package}
\seealso{
	mvmeta
}
