\name{hamling}
\alias{hamling}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Hamling's method
}
\description{
This function approximates the covariance matrix of log relative risks according to the method described by Hamling (2008).
}
\usage{
hamling(data)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{data}{
the dataset with the information needed to perform the algorithm.
}
}
\details{
No further details required.
}
\value{
A matrix reporting the pseudo-counts originated from the pubblished data.}
\references{
Hamling, J., Lee, P., Weitkunat, R.,  Ambuhl, M. (2008). Facilitating meta-analyses by deriving relative effect and 
precision estimates for alternative comparisons from a set of estimates presented by exposure level or disease category. 
Statistics in medicine, 27(7), 954-970.}
\author{
Alessio Crippa
}