% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{ir_ex}
\alias{ir_ex}
\title{Incidence-rate data on fiber intake and coronary heart disease risk}
\format{A data frame with 5 observations on the following 8 variables:
\tabular{ll}{
\code{dose} \tab assigned dose levels.\cr
\code{cases} \tab number of cases for each exposure level.\cr
\code{n} \tab total number of subjects for each exposure level.\cr
\code{adjrr} \tab adjusted incidence rate ratios for each exposure level.\cr
\code{lb} \tab lower bound for the confidence limits of the adjusted incidence rate ratios.\cr
\code{ub} \tab upper bound for the confidence limits of the adjusted incidence rate ratios.\cr
\code{logrr} \tab natural logarithm of adjusted incidence rate ratios.\cr
\code{se} \tab standard error for the logarithm of the adjusted incidence rate ratios.\cr
}}
\description{
The dataset reports the summarized dose-response results from incidence-rate data 
investigating the association between the long-term intake of dietary fiber and 
risk of coronary heart disease among women, first presented by Wolk et al. (1999)
}
\references{
Wolk, A., J. E. Manson, M. J. Stampfer, G. A. Colditz, F. Hu, F. E. Speizer, C. H. Hennekens, and W. C. Willett. 1999. 
Long-term intake of dietary fiber and decreased risk of coronary heart disease among women. Journal of the American Medical Association 281: 1998-2004.

Greenland, S.,  Longnecker, M. P. (1992). Methods for trend estimation from summarized dose-response data, with applications to meta-analysis.
American journal of epidemiology, 135(11), 1301-1309.
}
\author{
Alessio Crippa, <\email{alessio.crippa@ki.se}>
}
\keyword{data}
